requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
qrcode==7.3.1
Pillow==8.2.0
